Refactor camkes simulation tests to use CTest
CTest allows the tests to be better integrated with the project build
configuration. This makes maintaining and adding tests easier. It also
makes it possible to run the tests as part of a development workflow
with a near-instant incremental rebuild instead of having to configure
and build a full project from scratch.